+++ src/ft_realloc.c
@@ -6,16 +6,16 @@
 /*   By: glegendr <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/12 21:37:20 by glegendr          #+#    #+#             */
-/*   Updated: 2017/12/12 21:46:21 by glegendr         ###   ########.fr       */
+/*   Updated: 2017/12/19 18:30:46 by glegendr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		*ft_realloc(void *mem, size_t mem_size, size_t new_size)
+void		*ft_realloc(void *mem, int mem_size, int new_size)
 {
 	void *tmp;
-	
+
 	if ((tmp = malloc(new_size)) == NULL)
 		return (NULL);
 	if (mem == NULL)
